+++ changelog.md
@@ -23,6 +23,7 @@
 - `::->` Resolver is Monad
 
 - Basic skeleton for subscriptions. An actual way to use subscriptions is still work in progress.
+- inline Fragments
 
 ### Changed